Rewrite of Test cases and change of storage
Test cases were rewritten to use FileRow. Some tests will be fixed when file metadata amount is found.

Portion calculation has changed to be relative to current_root and not its immediate parent.

An algorithm for get_paint_order() has been made with plans to add a "layer" field to the Node struct.